+++ app/educator-os/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import RoleOSDashboard from "@/components/role-os/RoleOSDashboard";
+
+export default function EducatorOSPage() {
+  return <RoleOSDashboard role="educator" />;
+}